Add item and fluid null check method (#105)
* update dependencies

* ✨ feat(tag): Expand terracotta tag with stained clay variants

Updates the terracotta tag to include all stained hardened clay variants,
improving item grouping and compatibility. Removes unused stained_hardened_clay tag.

* ✨ feat(condition): Add TagEmptyCondition serializer

Registers the TagEmptyCondition serializer to allow checking for empty tags.

* ✨ feat(inventory): Add helper methods for inventory manipulation

Introduces new methods for setting player inventory slots, filling slots,
merging item stacks, and adding items to multiple slots.

* ✨ feat(inventory): Refine inventory helper methods

Improves `fillSlot` and `addToStack` for better handling of null or empty item stacks.
Adds `markDirty()` calls for inventory updates and ensures item stack copying.

* ✨ feat(condition): Change TagEmptyCondition name to forge

Renames the `tag_empty` condition to `forge:tag_empty` for better
compatibility and consistency with Forge's naming conventions.

* ✨ feat(recipe): Use Ingredient.valueFromJson for tag lookup

Replaces OreDictionary lookup with Ingredient.valueFromJson for tag resolution.
Handles null or empty matching stacks with a warning log.
Adds support for "amount" as a count key in recipes.

* ✨ feat(dependencies): Update JFMUY version

Updates the JFMUY dependency in dependencies.gradle to the latest version.
✨ feat(block): Add register method for block with item instance

Adds a new `register` method to `BlockAction` that accepts an `Item` instance
for block registration, improving flexibility.

* ✨ feat(network): Make getAction public

Changes the visibility of `getAction` method in `PacketCodec` from protected to public.

* ✨ feat(network): Make ICodecAction and ICodecRunnable public

Exposes `ICodecAction` and `ICodecRunnable` interfaces within `PacketCodec`
to public visibility.

* ✨ feat(network): Add registerCodecAction method

Adds a static `registerCodecAction` method to `PacketCodec` for registering codec actions.

* sa

* ✨ feat(network): Register codec action for DimPos

Adds a codec action to PacketCodec for encoding and decoding DimPos objects.

* ✨ feat(network): Move DimPos codec registration

Relocates the DimPos codec registration to the end of the PacketCodec
initialization block.

* ✨ feat(ingredient): Change ENERGY component type to Long

Updates the `ENERGY` `IngredientComponent` to use `Long` instead of `Integer`
for its generic type.

* ✨ feat(inventory): Add setters for slot position

Introduces `setSlotPosX` and `setSlotPosY` to modify slot display
positions, updating existing logic in `addSlotToContainer`.

* ✨ feat(ingredient): Use Long for ENERGY component

Updates the ENERGY ingredient component to use Long instead of Integer
for its type, allowing for larger energy values.

* ✨ feat(capabilities): Refactor DefaultCapabilityProvider

Simplifies the constructor and the getCapability method,
improving clarity and reducing redundant code.

* 🗑️ refactor(capabilities): Remove SerializableCapabilityProvider

* 🗑️ refactor(capabilities): Remove SerializableCapabilityProvider

* ✨ feat(gui): Add GuiButtonCheckbox component

Introduces a new checkbox button for GUI elements,
allowing for state toggling with visual feedback.

* 🐛 fix(network): Correctly register DimPos codec action

Changes `registerCodecAction` to `addCodedAction` and uses `codecActions.put`
directly for `DimPos` class registration.

* ✨ feat(recipe): Extract ItemStack variants logic to helpers

Moves the `getItemStackVariants` method from
`PrototypedIngredientAlternativesItemStackOredictionary` to
`ItemStackHelpers`.

* ✨ feat(glstate): Add GlStateManager utility methods

Introduces helper methods for OpenGL state management, including texture, alpha test, blend function, and color settings.

* ✨ feat(command): Add Enum argument type

Introduces a generic ArgumentType for enums, allowing for command
arguments that accept enum values with suggestions.

* ✨ feat(recipe): Add reusable input tracking to RecipeDefinition

Introduces a new field and logic to `RecipeDefinition` to track whether recipe inputs are reusable. This allows for more accurate consumption logic in crafting.

* ✨ feat(collection): Add quantity management to collapsed collections

Implements getQuantity and setQuantity methods in IngredientCollectionMutable
and IIngredientCollapsedCollectionMutable, allowing for quantity manipulation
of collapsed ingredient collections.

* ✨ feat(helper): Add ValueNotifierHelpers for NBT serialization

Introduces helper methods to ValueNotifierHelpers for setting and getting
various NBT types (NBTTagCompound, int, boolean, String, List<String>)
to simplify serialization and deserialization logic.

* ✨ feat(helper): Improve NBT string and list handling

Adds checks for null or empty strings when setting NBT string and list values.
Retrieves empty strings correctly when they are stored as spaces.

* ✨ refactor(item): Consolidate ItemStack empty checks

Replaces `isStackEmpty` with a new `isEmpty` method and updates related usages.
Also renames `getItemStackVariants` to `getVariants` for clarity.

* ✨ refactor(ItemStackHelpers): Improve null and empty checks

Adds null and empty checks to various ItemStack helper methods to prevent
NPEs and ensure correct behavior with empty stacks.

* ✨ refactor(ItemStackHelpers): Use EMPTY stack for null checks

Refactors various methods to use ItemStackHelpers.EMPTY instead of null checks,
improving code clarity and consistency.

* ✨ refactor(MinecraftHelpers): Remove redundant drop/clear inventory logic

Removes the logic for dropping and clearing inventory from BlockOK in preDestroyBlock,
as this functionality is now handled by BlockTile.

* ✨ refactor(helpers): Consolidate helper classes and imports

Moves FluidHelpers and EnergyHelpers to the helper package.
Removes ItemHelpers and merges its functionality into ItemStackHelpers.
Updates imports across the project to reflect these changes.

* ✨ refactor(ItemHelpers): Consolidate ItemStack helper methods

Replaces the `ItemStackHelpers` class with the new `ItemHelpers` class.
This change centralizes ItemStack-related utility methods for better organization and
maintainability across the project.

* ✨ refactor(FluidHelpers): Add isEmpty, copy, copyWithAmount and EMPTY

Introduces `isEmpty`, `copy`, `copyWithAmount` methods and a static `EMPTY` constant to `FluidHelpers` for consistent null and empty fluid stack handling.

* ✨ refactor(FluidHelpers): Use ItemHelpers.isEmpty for null checks

Replaces null checks with ItemHelpers.isEmpty for better code consistency and readability in FluidHelpers.

* optimizing images (#106)


* ✨ refactor(ItemHelpers): Remove redundant areStacksEqual method

Removes the `areStacksEqual` method from `ItemHelpers` as it is an alias for `areItemsEqual` and no longer needed.

---------

+++ src/main/java/ruiseki/commoncapabilities/api/ingredient/IngredientComponent.java
@@ -56,7 +56,7 @@ public final class IngredientComponent<T, M>
 
     public static IngredientComponent<ItemStack, Integer> ITEMSTACK = null;
     public static IngredientComponent<FluidStack, Integer> FLUIDSTACK = null;
-    public static IngredientComponent<Integer, Boolean> ENERGY = null;
+    public static IngredientComponent<Long, Boolean> ENERGY = null;
 
     @CapabilityInject(IIngredientComponentStorageHandler.class)
     private static Capability<IIngredientComponentStorageHandler> CAPABILITY_INGREDIENT_COMPONENT_STORAGE_HANDLER = null;
